+++ src/test/java/io/vertx/core/http/RequestOptionsTest.java
@@ -36,6 +36,8 @@ public class RequestOptionsTest {
     assertEquals(RequestOptions.DEFAULT_URI, options.getURI());
     assertEquals(RequestOptions.DEFAULT_FOLLOW_REDIRECTS, options.getFollowRedirects());
     assertEquals(RequestOptions.DEFAULT_TIMEOUT, options.getTimeout());
+    assertEquals(RequestOptions.DEFAULT_CONNECT_TIMEOUT, options.getConnectTimeout());
+    assertEquals(RequestOptions.DEFAULT_IDLE_TIMEOUT, options.getIdleTimeout());
   }
 
   @Test
@@ -63,6 +65,8 @@ public class RequestOptionsTest {
       .addHeader("foo", Arrays.asList("bar", "baz"));
     JsonObject expected = new JsonObject()
       .put("timeout", RequestOptions.DEFAULT_TIMEOUT)
+      .put("connectTimeout", RequestOptions.DEFAULT_CONNECT_TIMEOUT)
+      .put("idleTimeout", RequestOptions.DEFAULT_IDLE_TIMEOUT)
       .put("uri", RequestOptions.DEFAULT_URI)
       .put("method", "PUT")
       .put("port", 8443)
